Creates basic VTT file parser.
Accepts an input stream to allow as many kinds of sources as possible
(most if not all production code should be a file input stream, but test
code might want an input stream built from a string).

Parser is currently very faithful to the algorithm in the official spec,
but it isn't very intuitive. Saving this before trying a different
approach.

Still much to do here.