Tokenizes cue text.
Text was previously stored unprocessed in a raw cue. This converts it
into a list of tokens (either plain text, start tag, end tag, or
timestamp tag). Parsing these tokens is still a TODO because this is
already a pretty big change.

Requires importing Apache commons library for StringEscapeUtils to
unescape HTML character references.

Modifies Cue to contain a single VttNode as the payload, because the
VttLine solution would have made inconsistent results when the cue text
was all unformatted text on multiple lines vs. a line break within an
HTML tag. Also, this would have changed the "DOM" structure, which could
change the way the CSS rules match the text down the line.