Implement CSS Background shorthand.
Refactor cascading background objects to include "background" in the
name to disambiguate from plain Color type.
There is a new shorthand interface to ensure shorthand properties
implement all required functionality. In the future this is planned to
be something like getting the initial value as Css property
(essentially, return itself instead of a resolved property like it does
currently).