Implement methods on CSS property interfaces to allow style cascading.
Requires support from each property (gain a new cascadeFrom() method)
and a way for property values to convert their value to a computed
value. Most of these are a TODO because of how many properties there
are, but some trivial ones are implemented.

As a result, all types previously used in CssProperty now must implement
a new CssValue interface. FontFamilyList and LineTypeSet have been
introduced to satisfy generic type constraints.

Eventually there will be a visitor to perform the cascade, but this
change is already big enough.