Adds basic support for font cascading.
This has been in progress for a while and there are a lot of TODOs
remaining, but cascading fonts is pretty difficult. Color and length
measurement need to be cascaded properly (and length is the biggest TODO
here).

- Changes method signature of computeValue to take a
CascadeValueReference, created to make this all easier. It is a
reference to a parent and self value so a CssValue itself can decide how
to cascade.
   - A bunch of refactoring to support this
- In the future, font will be needed for measurement (em unit), so this
is causing me to re-think how font is represented. It will likely use
the Font class that has been sitting around until now, to be more
consistent with the use of Color in CascadeValueReference.
   - The shorthand classes may be converted to be used for updating
values.